Log mouse clicks. Closes #6.
Mouse cursor logging is now done using pygame instead of Xlib.
Because of this, coordinates are now logged in relation to
the top left corner of the pygame display window instead
of in relation to the corners of the whole screen.

(Of course mouse clicks, as events, do not follow the same
framerate as mouse position logging. Despite this, mouseclick
logs are written among the mouse position logs.)